* Move provisioning function into its own file
* Catch partner provision command failures
